全面修复代码安全、逻辑缺陷及代码质量问题
安全修复:
- JWT密钥从硬编码改为application.yml配置注入
- 数据库密码使用环境变量占位符(默认值仅本地开发)
- 评委默认密码改为配置项注入
- BCryptPasswordEncoder统一为Spring Bean

逻辑修复:
- TeamSqlProvider comId参数类型与Mapper对齐(Integer)
- TeamServiceImpl中isCaptain字段Integer拆箱改用Objects.equals防NPE
- updateInstructors删除指导老师时检查其他队伍关联再置null
- JwtUtils升级至jjwt 0.13新API(SecretKey)

代码质量:
- IntructorMapper重命名为InstructorMapper(修正拼写)
- AOP包名改为小写aop,类名改为LogAspect避免框架类冲突
- UserMapper.selectByUserCode参数改为String
- ScoreMapper.updateById移除冗余SET id子句
- 添加RoleConstants常量类消除魔法数字
- LoginController增加输入空值校验
- ScoreController增加分数0-100范围校验
- Controller变量命名统一为camelCase

+++ sast-woc-project/src/main/java/org/example/provider/TeamSqlProvider.java
@@ -5,7 +5,7 @@ import org.apache.ibatis.jdbc.SQL;
 public class TeamSqlProvider {
 
     public String selectByComIdAndAcademy(@Param("academyId") Long academyId,
-                                          @Param("comId") Long comId) {
+                                          @Param("comId") Integer comId) {
         return new SQL() {{
             SELECT("t.*");
             FROM("team t");
